+++ mixins/externals.js
@@ -13,12 +13,14 @@ module.exports = {
 		const htmlPluginInstance = helper.getPluginByName(config, 'HtmlWebpackPlugin');
 		const webOSMetaPluginInstance = helper.getPluginByName(config, 'WebOSMetaPlugin');
 
-		const libraries = ['@enact', 'react', 'react-dom', 'ilib'];
+		const libraries = ['@enact', 'react', 'react-dom', 'react-dom/client', 'react-dom/server', 'ilib'];
 
 		const app = packageRoot();
 		if (
 			app.meta.name.startsWith('@enact/') &&
-			(fs.existsSync(path.join(app.path, 'ThemeDecorator')) || app.meta.name === '@enact/i18n')
+			(fs.existsSync(path.join(app.path, 'MoonstoneDecorator')) ||
+				fs.existsSync(path.join(app.path, 'ThemeDecorator')) ||
+				app.meta.name === '@enact/i18n')
 		) {
 			libraries.push('.');
 		}