+++ plugins/dll/EnactFrameworkRefPlugin.js
@@ -86,13 +86,21 @@ class EnactFrameworkRefPlugin {
 	constructor(options = {}) {
 		this.options = options;
 		this.options.name = this.options.name || 'enact_framework';
-		this.options.libraries = this.options.libraries || ['@enact', 'react', 'react-dom', 'ilib'];
+		this.options.libraries = this.options.libraries || [
+			'@enact',
+			'react',
+			'react-dom',
+			'react-dom/client',
+			'react-dom/server',
+			'ilib'
+		];
 		this.options.ignore = this.options.ignore || [
 			'@enact/dev-utils',
 			'@enact/storybook-utils',
 			'@enact/ui-test-utils',
 			'@enact/screenshot-test-utils',
-			'readable-stream'
+			'readable-stream', // ignore for screenshot test build
+			'react-is' // ignore for ui test build
 		];
 		this.options.external = this.options.external || {};
 		this.options.external.publicPath =